% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{gg_translate}
\alias{gg_translate}
\title{Translate all text and factor labels in ggplot2 or patchwork plots}
\usage{
gg_translate(plot, trans)
}
\arguments{
\item{plot}{A `ggplot` or `patchwork` plot object to translate.}

\item{trans}{A named character vector of translations.
Names correspond to source-language terms and values
to target-language translations.}
}
\value{
A translated plot object of the same type.
}
\description{
`gg_translate()` provides a generic way to translate plot texts,
axis labels, legend titles, legend entries, and facet strips
in an existing ggplot2 or patchwork plot using a named translation
vector.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(wt, mpg, colour = factor(cyl))) +
  geom_point() +
  labs(title = "Car efficiency", colour = "Cylinders")

trans <- c("Car efficiency" = "Autotehokkuus",
           "Cylinders" = "Sylinterit",
           "4" = "4-syl.",
           "6" = "6-syl.",
           "8" = "8-syl.")

gg_translate(p, trans)

}
