% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggcustom_pal}
\alias{ggcustom_pal}
\title{Color palettes in ggcustom package.}
\usage{
ggcustom_pal(n = NULL, name, last_grey = FALSE)
}
\arguments{
\item{n}{A number of colors needed. NULL (default) returns all available colors.}

\item{name}{A name of the palette. See \code{names(ggcustom_palettes)}.}

\item{last_grey}{Logical, if TRUE, the last color will be replaced with "grey75".}
}
\description{
Color palettes in the ggcustom package (discrete).
}
\details{
Colors in palettes are recycled if needed (with warning).

Available palettes are: vm.
}
\examples{
scales::show_col(ggcustom_pal(n = 6, name = "vm"))
scales::show_col(ggcustom_pal(n = 6, name = "fpb"))
}
