% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{vm_pal}
\alias{vm_pal}
\alias{vm_pal_lg}
\title{VM Color Palette}
\usage{
vm_pal(n = NULL)

vm_pal_lg(n = NULL)
}
\arguments{
\item{n}{A number of colors needed.}
}
\description{
Color palette for VM (discrete).
}
\details{
Provides the VM color palette with colors recycled if needed (with warning).
}
\section{Functions}{
\itemize{
\item \code{vm_pal_lg()}: VM palette with the last color grey.

}}
\examples{
scales::show_col(vm_pal(6))
}
