% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_vm}
\alias{scale_fill_vm}
\alias{scale_colour_vm}
\alias{scale_color_vm}
\alias{scale_colour_vm_lg}
\alias{scale_color_vm_lg}
\title{VM Color and Fill Scale}
\usage{
scale_fill_vm(...)

scale_colour_vm(...)

scale_color_vm(...)

scale_colour_vm_lg(...)

scale_color_vm_lg(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}.}
}
\description{
Discrete VM color scale. Colors from \code{\link{vm_pal}}.
}
\examples{
 library(ggplot2)
 dsamp <- diamonds[sample(nrow(diamonds), 1000), ]
 (d <- ggplot(dsamp, aes(carat, price, colour = clarity)) + geom_point())

 d + scale_colour_vm()
}
