% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_gg}
\alias{set_gg}
\alias{unset_gg}
\title{Set Theme and Scales}
\usage{
set_gg(theme, palette)

unset_gg()
}
\arguments{
\item{theme}{A ggplot2 theme.}

\item{palette}{A palette function or the name of a ggcustom palette.
See \code{names(ggcustom_palettes)}.}
}
\description{
Set a custom theme and scales using ggcustom. Updates `theme` and discrete
color and fill scales with the selected palette.
}
\section{Functions}{
\itemize{
\item \code{unset_gg()}: Restores the original theme and scales before `set_gg()` was called.

}}
