% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_vm}
\alias{set_vm}
\title{Set VM Theme and Palette}
\usage{
set_vm()
}
\description{
This is a shortcut function to set the VM theme and palette for `ggplot2` visualizations.
It applies the \code{\link{theme_vm}} and sets the default discrete color and fill scales
to use the VM palette.
}
\examples{
library(ggplot2)
library(ggcustom)

# Example dataset
data <- data.frame(
  category = rep(letters[1:4], each = 3),
  value = c(4, 3, 5, 2, 6, 7, 3, 5, 4, 6, 2, 7),
  group = rep(1:3, times = 4)
)

# Define a ggplot object
p <- ggplot(data, aes(x = category, y = value, colour = factor(group), group = group)) +
  geom_line(size = 1.2) +
  geom_point(size = 3)

# Apply the VM theme and palette
set_vm()
p

# Reset to original theme and scales
unset_gg()
}
