% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fpb}
\alias{theme_fpb}
\title{Theme FPB}
\usage{
theme_fpb(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{a font size}

\item{base_family}{a font}
}
\description{
Theme FPB. Based on \code{\link[ggplot2]{theme_bw}}. Sets discrete palette defaults
via \code{ggplot2} theme entries and geom defaults from \code{\link{fpb_pal}}
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p
p + theme_fpb()
}
