% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_vm}
\alias{theme_vm}
\title{Theme VM}
\usage{
theme_vm(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{a font size}

\item{base_family}{a font}
}
\description{
Theme VM. Based on \code{\link[ggplot2]{theme_bw}}. Sets discrete palette defaults
via \code{ggplot2} theme entries and geom defaults from \code{\link{vm_pal}}
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p
p + theme_vm()
}
